+++ CHANGELOG.md
@@ -7,9 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added config key `public`
+- Added module `cogs.debug`
+
 ### Changed
 
 - Renamed `VERSION` in module `const` to `PROJECT_VERSION`
+
+### Removed
+
+- Removed config key `auto-remove`
 
 ## [0.1.0] - 2026-04-18